requirements.txt
pillow
openai
azure-ai-vision-imageanalysis
spacy
spacy-wordnet
streamlit
